requirements.txt
absl-py==1.4.0
adagio==0.2.4
aenum==3.1.11
aiohttp==3.8.1
aiosignal==1.3.1
alembic==1.8.1
altair==4.2.0
antlr4-python3-runtime==4.11.1
anyio==3.6.2
appdirs==1.4.4
apsw==3.34.0.post1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
awswrangler==2.17.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
blis==0.7.9
bokeh==2.4.3
Boruta==0.3
boto3==1.26.10
botocore==1.29.10
cachetools==5.2.0
catalogue==1.0.2
category-encoders==2.5.1.post0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorcet==3.0.1
colorlover==0.3.0
commonmark==0.9.1
configparser==5.3.0
cryptography==3.3.2
css-parser==1.0.6
cssselect==1.1.0
cssutils==1.0.2
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
dask==2023.1.0
databricks-cli==0.17.3
datashader==0.14.3
datashape==0.5.2
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
deepchecks==0.9.2
defusedxml==0.7.1
distro-info==1.0
Django==3.1.14
dnspython==2.0.0
docker==6.0.1
emoji==2.2.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.2.0
fairlearn==0.8.0
fastjsonschema==2.16.2
feedparser==5.2.1
fitter==1.5.2
Flask==2.2.2
flatbuffers==23.3.3
fonttools==4.38.0
fpdf==1.7.2
fpdf2==2.6.0
frozendict==2.3.6
frozenlist==1.3.3
fs==2.4.16
fsspec==2022.11.0
fugue==0.7.3
fugue-sql-antlr==0.1.1
funcy==1.17
future==0.18.2
gast==0.4.0
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.1
gremlinpython==3.6.1
grpcio==1.53.0
gunicorn==20.1.0
gyp==0.1
h11==0.14.0
h5py==3.8.0
holoviews==1.15.4
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.1
htmlmin==0.1.12
idna==3.4
ifaddr==0.1.7
ImageHash==4.3.1
imageio==2.24.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpath-ng==1.5.3
jsonpickle==3.0.0
jsonschema==4.17.0
jupyter-server==1.23.3
jupyter_client==7.4.5
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kds==0.1.3
keras==2.12.0
kiwisolver==1.4.4
kmodes==0.12.2
libclang==16.0.0
lightgbm==3.3.3
llvmlite==0.36.0
locket==1.0.0
lxml==4.9.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mechanize==0.4.5
missingno==0.5.1
missingpy==0.2.0
mistune==2.0.4
ml-dtypes==0.0.4
mlflow==1.30.0
mlxtend==0.19.0
msgpack==1.0.0
multidict==6.0.2
multimethod==1.9
multipledispatch==0.6.0
multitasking==0.0.11
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.9
networkx==2.8.8
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.53.0
numexpr==2.8.4
numpy==1.23.4
numpy-financial==1.0.0
oauthlib==3.2.2
olefile==0.46
openai==0.27.0
openpyxl==3.0.10
opensearch-py==2.0.0
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandas-profiling==3.6.6
pandocfilters==1.5.0
panel==0.14.2
param==1.12.3
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pexpect==4.8.0
pg8000==1.29.3
phik==0.12.2
pickleshare==0.7.5
Pillow==9.3.0
plac==1.1.3
platformdirs==2.5.3
plotly==5.11.0
ply==3.11
pmdarima==2.0.3
preshed==3.0.8
progressbar2==4.2.0
prometheus-client==0.15.0
prometheus-flask-exporter==0.21.0
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
py7zr==0.0.0
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyathena==2.14.0
pycaret==2.3.10
pyChatGPT==0.0.7
pychm==0.8.6
pycparser==2.21
pycryptodomex==3.9.7
pyct==0.4.8
pycurl==7.43.0.6
pydantic==1.10.2
pydeck==0.8.0
pygame==2.1.2
Pygments==2.13.0
PyGObject==3.38.0
PyJWT==2.6.0
pyLDAvis==3.3.1
Pympler==1.0.1
PyMySQL==1.0.2
pynndescent==0.5.8
PyNomaly==0.3.3
pyod==1.0.6
pypandoc==1.10
pyparsing==3.0.9
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.2
pyrsistent==0.19.2
PySocks==1.7.1
pyspark==3.3.1
python-apt==2.2.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-utils==3.4.5
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==23.2.1
qpd==0.3.3
querystring-parser==1.2.4
redshift-connector==2.0.909
regex==2022.10.31
repoze.lru==0.7
requests==2.28.1
requests-aws4auth==1.1.2
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.6.0
Routes==2.5.1
rsa==4.9
s3transfer==0.6.0
schedule==1.1.0
scikit-image==0.19.3
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.10.1
scramp==1.4.4
seaborn==0.12.1
selenium==4.7.2
semver==2.13.0
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==2.3.8
SQLAlchemy==1.4.44
sqlparse==0.4.3
srsly==1.0.6
stack-data==0.6.1
statsmodels==0.13.5
streamlit==1.14.1
streamlit-elements==0.1.0
streamlit-pandas-profiling==0.1.3
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
terminado==0.17.1
textblob==0.17.1
texttable==1.6.3
thinc==7.4.6
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
triad==0.7.0
trio==0.22.0
trio-websocket==0.9.2
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
umap-learn==0.5.3
unattended-upgrades==0.1
urllib3==1.26.12
uuid==1.30
validators==0.20.0
visions==0.7.5
wasabi==0.10.1
watchdog==2.1.9
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wordcloud==1.8.2.2
wrapt==1.14.1
wsproto==1.2.0
xarray==2023.1.0
xdg==5
xgboost==1.7.4
yarl==1.8.1
ydata-profiling==4.0.0
yellowbrick==1.2.1
yfinance==0.2.14
zeroconf==0.26.1
zipp==3.10.0
